+++ ShadowPluginLoader.WinUI/AbstractPluginLoader.Change.cs
@@ -1,5 +1,9 @@
-﻿using System.Threading.Tasks;
+﻿using System.IO;
+using System.Threading.Tasks;
+using ShadowPluginLoader.WinUI.Args;
 using ShadowPluginLoader.WinUI.Checkers;
+using ShadowPluginLoader.WinUI.Enums;
+using ShadowPluginLoader.WinUI.Exceptions;
 
 namespace ShadowPluginLoader.WinUI;
 
@@ -20,5 +24,20 @@ public abstract partial class AbstractPluginLoader<TMeta, TAPlugin>
     {
         await RemoveChecker.CheckRemoveAsync();
         await UpgradeChecker.CheckUpgradeAsync();
+    }
+
+
+    /// <summary>
+    /// <inheritdoc />
+    /// </summary>
+    public Task RemovePlugin(string id)
+    {
+        var plugin = GetPlugin(id);
+        if (plugin == null) throw new PluginRemoveException($"{id} Plugin Not Found");
+        var dir = Path.GetDirectoryName(plugin.GetType().Assembly.Location);
+        if (dir == null) throw new PluginRemoveException($"{id} Plugin Path Not Found");
+        RemoveChecker.PlanRemove(id, dir);
+        PluginEventService.InvokePluginPlanRemove(this, new PluginEventArgs(id, PluginStatus.PlanRemove));
+        return Task.CompletedTask;
     }
 }
