+++ ShadowPluginLoader.WinUI/Interfaces/IPluginLoader.Change.cs
@@ -4,10 +4,15 @@ namespace ShadowPluginLoader.WinUI.Interfaces;
 
 public partial interface IPluginLoader<TMeta, TAPlugin>
 {
-
     /// <summary>
     /// Checked for updates and removed plugins
     /// </summary>
     /// <returns></returns>
     Task CheckUpgradeAndRemoveAsync();
+
+    /// <summary>
+    /// Remove Plugin
+    /// </summary>
+    /// <param name="id">Plugin Id</param>
+    Task RemovePlugin(string id);
 }
